Modifiable Entity Extra Data
Allowed user to modify entity extra data
extra1 is max health (short) and extra2 is type (byte)

+++ src/javajesus/entities/Mob.java
@@ -988,8 +988,8 @@ public abstract class Mob extends Entity implements Damageable, Hideable, Skills
 	/**
 	 * Changes the mob's max health
 	 */
-	public void setMaxHealth(int health) {
-		maxHealth = (short) health;
+	public void setMaxHealth(short health) {
+		maxHealth = health;
 	}
 
 	/**
